Add files via upload
These are the files for the assignment from last class on Thursday. I edited the code to run only the first column and took pictures of the original graphs along with the graphs after editing the code. The code in part a was unchanged, it worked just fine as is. The graphs labled "Figure" 1-6 are the post-editing graph photos.